Create views for most recent spark output data (Recidiviz/recidiviz-data#24216)
## Description of the change

This PR fixes an overlooked issue in Spark, which is that there is no
distinction between different runs of a simulation when pulling Spark BQ
output metrics into a Google Sheets pivot table. Instead of instituting
a protocol of changing the simulation tag every time one runs a
simulation, this PR allows folks to keep using the same simulation tag
by adding views to the spark output dataset that pull only the most
recent data for each simulation tag. These views are then queried by
pivot tables instead of the originals.

I had problems testing this / the pre-commit would yell at me and I'm
pretty sure it's because I'm trying to affect parts of recidiviz's
codebase I don't have access to as a contractor. @agaidus would you be
able to take a look at this and make sure I implemented what we talked
about correctly? Thanks so much!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Paco Poler <pacopoler@Pacos-MacBook-Pro.local>
Co-authored-by: Andrew Gaidus <andrewg@recidiviz.org>
GitOrigin-RevId: e1b4e976b1613b16cc7374f72794710431a46bce