[Admin Panel] Add API for raw data watermarks (Recidiviz/recidiviz-data#24321)
## Description of the change

This adds an endpoint to query the raw data watermarks for the
operations DB, for display in the admin panel.

This just gets the max timestamp for each raw data table from the
database. There is no way to tell what the
latest pipeline was based on the table, so we rely on the fact that
watermarks should only increase to infer
that the maximum watermark was used by the most recent pipeline. If for
some reason a raw data table is no
longer used at all, but there are still rows, it will show up on the
page until the rows are invalidated

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to Recidiviz/recidiviz-data#24081

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d067c7f4d36fccdbe1d390e88b2129d29788c7fe