[Justice Counts] [2/n] Create new endpoint to copy over metric config to child agencies (Recidiviz/recidiviz-data#24402)
## Description of the change

New endpoint for the admin panel API that will copy over the metric
config from a super agency to a child agency. This uses the same code as
the script but removes the logging and dry-run option.

## Related issues

Related to https://github.com/Recidiviz/recidiviz-data/issues/24389

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c1c8f895b7949cbb2f1f8e452933fbf99b81c0a1