[Ingest on Dataflow] Add admin panel API for state dataset row counts (Recidiviz/recidiviz-data#24473)
## Description of the change

This does roughly the same thing as Recidiviz/recidiviz-data#24429, but for state instead of
ingest view results. I pushed the code for getting the row counts of all
tables in a dataset into the big query client, so that it could be
shared for both of these endpoints and in case it is useful for any
other BQ datasets in the future. I left this as a separate endpoint,
instead of adding it into the ingest view results endpoint, to keep the
json return structure simple and allow us to make the calls in parallel
from the frontend instead of needing to spin up multiple threads in the
backend.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

towards Recidiviz/recidiviz-data#24081

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 40fbfe99a971ea0c50266a29929a9352e86f3fd6