[schema change] Update supervision_officer_attribute_sessions to use new field (Recidiviz/recidiviz-data#24496)
## Description of the change

Since Recidiviz/recidiviz-data#24269 has been deployed to prod and staging, we can change all
downstream views to use the new `caseload_type` field instead of
`state_staff_specialized_caseload_type`. This will allow that field to
have a `GENERAL` value, so in this change I updated which field is
pulled into `supervision_officer_attribute_sessions` and changed the
name of the field in that view to reflect that this will be populated
for all supervision officers, not just those with a specialized caseload
type.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Anna Geiduschek <anna@recidiviz.org>
GitOrigin-RevId: e3b6fea8ca8ec14e61137fcf300945e5d049d39a