Remove `secret_prefix_override` parameter from our DB connection code (Recidiviz/recidiviz-data#24581)
## Description of the change

In https://github.com/Recidiviz/recidiviz-data/pull/23214, I added this
`secret_prefix` parameter to a fair bit of our DB connection code. The
reason this was needed is because we currently have an assumption in our
code that there is a 1:1 mapping between SchemaType and DB instance URL.
In the method `_secret_manager_prefix_for_type`, we use the SchemaType
to determine what secrets to look at to determine the instance URL.

The Justice Counts SchemaType is now used in two DB instances, one in
recidiviz-* GCP projects and one in the justice-counts-* GCP projects.
Fora temporary period of time, we needed to make it possible for the
recidiviz-* GCP projects to connect to both DB instances, and so we
needed two sets of secrets, one for each instance. The secret prefix
override allowed us to control which secrets we looked at and therefore
which instance we connected to.

Now, we have migrated all of JC infra to the new GCP projects, so the
recidiviz-* GCP projects only need to connect to one DB instance, and
the justice-counts-* GCP projects connect to the other one. Thus, each
project just needs to maintain one set of secrets, and so we can now
remove this override.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#23253

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 2075a88c90ea3882e6994114764364d55b338039