[Justice Counts] Overwrite CSG roles (Recidiviz/recidiviz-data#24607)
## Description of the change

This PR adds some additional logging to the `add_csg_old_agencies.py`
script. Additionally, we can now pass in `preserve_role` as an argument,
which determines whether or not to overwrite previously assigned
user/agency roles.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Michelle Orden <michelleorden@michelles-mbp.lan>
Co-authored-by: Michelle Orden <michelleorden@Michelles-MacBook-Pro.local>
GitOrigin-RevId: 2815efba9d5515e59732a9247df2cd984fe15eec