Revert "remove secret prefix override" (Recidiviz/recidiviz-data#24617)
## Description of the change

I messed up with https://github.com/Recidiviz/recidiviz-data/pull/24581.
When that PR got deployed to staging, the admin panel started talking to
the old JC database again. This change was NOT safe to make, because the
admin panel still needs to connect to the new JC database, so it is
STILL the case that recidiviz-* projects need to be able to talk to both
JC databases. When we migrate the admin panel to our own app (which is
in the works), THEN the change will be safe.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b9fe9f8432bb24e9841c39ae00965c15df4ad29a