[US_ME][JII] Script to generate ME SCCP eligibility data (Recidiviz/recidiviz-data#24099)
## Description of the change

This is a script meant to be run manually from the command line (for
now, at least!) for generating data for JII in Maine explaining their
SCCP eligibility. The script queries a specified BQ view (which must
have certain columns), parses and massages the data, and then outputs a
spreadsheet that can easily be used to autofill Google Docs templates
via an AppsScript that just does some basic find/replace.
[Here](https://docs.google.com/spreadsheets/d/1iO5rHaVw6XQRq3uzWwjC8hP20TJZ0YUgrjoMG9bHRuA/edit#gid=428666279)
is an example of a spreadsheet it generated, and
[here](https://script.google.com/u/0/home/projects/1uPNgDFGyBfdJZq8F_5OOC_lxXUe8s-Ec0qLMKLa2z0ZLwXbkyvio9yY3/edit)
is the AppsScript that takes the spreadsheet and autofills the Google
Docs and generates the pdfs.

As a followup, I'd like to move the actual autofilling of templates into
this script, so the process is 1-step rather than 2.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e80332d8c141cfcb3a4a51309dbab77806e2de16