[Testing] Disallow source table datasets in DEPLOYED_DATASETS_THAT_HAVE_EVER_BEEN_MANAGED (Recidiviz/recidiviz-data#24625)
## Description of the change

We had an incident this week where we deleted the
`spark_public_output_data` dataset because that dataset was accidentally
added to DEPLOYED_DATASETS_THAT_HAVE_EVER_BEEN_MANAGED, a list that
should only contain datasets that do / did contain views and their
materialized tables.

This PR adds a test that would have caught this error.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

See this Slack thread:
https://recidiviz.slack.com/archives/C028X32LRH7/p1697060477151169

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e6d462a832734fa368007f08c849ebb8da976099